requirements.txt
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.26.2
torch==2.1.2
torchvision==0.16.2
paddlepaddle==2.5.2
pyside6==6.6.1
matplotlib